requirements.txt
pytest
httpx
fastapi
uvicorn[standard]
scikit-learn
xgboost
pandas
requests
beautifulsoup4
openai
catboost
joblib
torch
# Database and async dependencies
supabase
python-dotenv
# Async and concurrency
aiohttp
asyncio-throttle
tenacity
psutil
# Connection pooling
aiopg
asyncpg